+++ src/main/java/com/mycompany/myapp/config/Constants.java
@@ -5,13 +5,13 @@ package com.mycompany.myapp.config;
  */
 public final class Constants {
 
-    // Regex for acceptable logins
-    public static final String LOGIN_REGEX = "^[_.@A-Za-z0-9-]*$";
-    
-    public static final String SYSTEM_ACCOUNT = "system";
-    public static final String DEFAULT_LANGUAGE = "ko";
-    public static final String ANONYMOUS_USER = "anonymoususer";
+	// Regex for acceptable logins
+	public static final String DEFAULT_LANGUAGE = "ko";
 
-    private Constants() {
-    }
+	public static final String SYSTEM_ACCOUNT = "system";
+	public static final String ANONYMOUS_USER = "anonymoususer";
+	public static final String LOGIN_REGEX = "^[_.@A-Za-z0-9-]*$";
+
+	private Constants() {
+	}
 }
