+++ src/main/java/com/mycompany/myapp/config/DatabaseConfiguration.java
@@ -29,21 +29,15 @@ public class DatabaseConfiguration {
         this.env = env;
     }
 
-    /**
-     * Open the TCP port for the H2 database, so it is available remotely.
-     *
-     * @return the H2 database TCP server.
-     * @throws SQLException if the server failed to start.
-     */
     @Bean(initMethod = "start", destroyMethod = "stop")
     @Profile(JHipsterConstants.SPRING_PROFILE_DEVELOPMENT)
     public Object h2TCPServer() throws SQLException {
-        String port = getValidPortForH2();
+        String port = getValidPortForH2Databse();
         log.debug("H2 database is available on port {}", port);
         return H2ConfigurationHelper.createServer(port);
     }
 
-    private String getValidPortForH2() {
+    private String getValidPortForH2Databse() {
         int port = Integer.parseInt(env.getProperty("server.port"));
         if (port < 10000) {
             port = 10000 + port;
