+++ src/main/java/com/mycompany/myapp/config/database/CloudDatabaseConfiguration2.java
@@ -11,18 +11,16 @@ import javax.sql.DataSource;
 import org.springframework.boot.context.properties.ConfigurationProperties;
 
 
+@Slf4j
 @Configuration
 @Profile(JHipsterConstants.SPRING_PROFILE_CLOUD)
 public class CloudDatabaseConfiguration extends AbstractCloudConfig {
-
-    private final Logger log = LoggerFactory.getLogger(CloudDatabaseConfiguration.class);
-    
     private static final String CLOUD_CONFIGURATION_HIKARI_PREFIX = "spring.datasource.hikari";
 
     @Bean
     @ConfigurationProperties(CLOUD_CONFIGURATION_HIKARI_PREFIX)
     public DataSource dataSource() {
-        log.info("Configuring JDBC datasource from a cloud provider");
+//        log.info("Configuring JDBC datasource from a cloud provider");
         return connectionFactory().dataSource();
     }
 }
